Updates: added multi-plotting & lettering capabilities
I added the ability to plot multiple plots and dimplots, still all with a single line of code!
I also added a capability into DBDimPlot for annotating dots with letters and set it to do this by default when the number of groups gets above 8.  In order to turn this off, if wanted, simply use do.letter = F.
I altered some of the default plot or axes titling.